backend/requirements.txt
Flask
flask_cors
PyMuPDF
python-docx
